tsconfig konfiguriert
