+++ frontend/src/components/Navigation.tsx
@@ -5,34 +5,52 @@ import { usePathname } from 'next/navigation';
 
 export default function Navigation() {
   const pathname = usePathname();
+
   return (
-    <nav style={{ padding: '10px', background: '#333', color: '#fff' }}>
+    <nav
+      style={{
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'space-between',
+        padding: '10px',
+        background: '#333',
+        color: '#fff',
+      }}
+    >
+      <div style={{ display: 'flex', gap: '15px' }}>
+        <Link href="/" style={{ color: pathname === '/' ? 'gray' : '#fff' }}>
+          Home
+        </Link>
+        <Link
+          href="/dummy"
+          style={{ color: pathname === '/orgUnit' ? 'gray' : '#fff' }}
+        >
+          Organisationseinheiten_test
+        </Link>
+        <Link
+          href="/organisationseinheiten"
+          style={{
+            marginRight: '15px',
+            color: pathname === '/organisationseinheiten' ? 'gray' : '#fff',
+          }}
+        >
+          Organisationseinheiten
+        </Link>
+        <Link
+          href="/rollen"
+          style={{
+            marginRight: '15px',
+            color: pathname === '/rollen' ? 'gray' : '#fff',
+          }}
+        >
+          Ermittle Rollen
+        </Link>
+      </div>
       <Link
-        href="/"
-        style={{
-          marginRight: '15px',
-          color: pathname === '/' ? 'gray' : '#fff',
-        }}
+        href="/login"
+        style={{ color: pathname === '/login' ? 'gray' : '#fff' }}
       >
-        Home
-      </Link>
-      <Link
-        href="/organisationseinheiten"
-        style={{
-          marginRight: '15px',
-          color: pathname === '/organisationseinheiten' ? 'gray' : '#fff',
-        }}
-      >
-        Organisationseinheiten
-      </Link>
-      <Link
-        href="/rollen"
-        style={{
-          marginRight: '15px',
-          color: pathname === '/rollen' ? 'gray' : '#fff',
-        }}
-      >
-        Ermittle Rollen
+        Login
       </Link>
     </nav>
   );
